converted helper functions into member functions of the outcome and added lots more tests
